When missing core files on Linux, show the current core file size limit.
A limit of 0 seems to be the default on some systems.

Only doing this check on Linux, because Mac falls back on the system crash reporter instead.

Only doing this check when actually looking for a core file, because I don't know how to tell ahead of time whether you're using breakpad or asan. Although doing this check up front anyway might not be a terrible idea, since those tools could fail and leave you with(out) a core dump anyway.